+++ internal/qbot/qbot.go
@@ -298,12 +298,19 @@ func (q *QBot) messageHandler(_ *discordgo.Session, m *discordgo.MessageCreate) 
 			return q.handleMoreTime(cmd)
 		case `!submitwave`, `!submitwaves`, `!wave`, `!waves`:
 			return q.handleSubmitWave(cmd)
-		case `!leaderboard`:
+		case `!leaderboard`, `!lb`:
 			return q.handleLeaderboard(cmd, false)
 		case `!history`:
 			return q.handleHistory(cmd)
 		case `!deverror`:
 			return q.handleDevError(cmd)
+		case `!loser`:
+			if len(cmd.Args) != 1 {
+				q.mustPost(cmd.Message.ChannelID, "Usage: `!loser <name>`")
+				return nil
+			}
+			q.congratulateLoser(cmd.Message.ChannelID, cmd.Args[0])
+			return nil
 		default:
 			q.mustPost(m.ChannelID, fmt.Sprintf("unknown command (use `!help` for available commands): `%s`", cmd.Command))
 			return nil